+++ 1-foundation/practical-logger-node/src/index.ts
@@ -1,23 +1,16 @@
-import { Logger, LoggerCreationParams } from '@offirmo/practical-logger-types'
+import { Logger, LoggerCreationParams, BaseSinkOptions } from '@offirmo/practical-logger-types'
 import { createLogger as createLoggerCore } from '@offirmo/practical-logger-core'
 
-const sink = (() => {
-	try {
-		// this version is for the author's own usage
-		// as it uses unpublished modules.
-		return require('./sinks/private').default
-	}
-	catch {
-		return require('./sinks/public').default
-	}
-})()
-
+import { SinkOptions } from './types'
+import createSink from './sinks/to-console'
 
 const ORIGINAL_CONSOLE = console
 
-function createLogger(p: Readonly<LoggerCreationParams>): Logger {
+
+
+function createLogger(p: Readonly<LoggerCreationParams<SinkOptions>> = {}): Logger {
 	return {
-		...createLoggerCore(p, sink),
+		...createLoggerCore(p, p.sinkOptions?.sink || createSink(p.sinkOptions)),
 		group(groupTitle?: string): void {
 			ORIGINAL_CONSOLE.group(groupTitle)
 		},